renaming entire application
