--- src/AuthApplication/azds.yaml
@@ -1,42 +1,0 @@
-kind: helm-release
-apiVersion: 1.1
-build:
-  context: .
-  dockerfile: Dockerfile
-install:
-  chart: charts/authapplication
-  values:
-  - values.dev.yaml?
-  - secrets.dev.yaml?
-  set:
-    replicaCount: 1
-    image:
-      repository: acmnotesregistry.azurecr.io/authapplication
-      tag: $(tag)
-      pullPolicy: Never
-    ingress:
-      annotations:
-        kubernetes.io/ingress.class: traefik-azds
-      hosts:
-      # This expands to form the service's public URL: [space.s.][rootSpace.]authapplication.<random suffix>.<region>.azds.io
-      # Customize the public URL by changing the 'authapplication' text between the $(rootSpacePrefix) and $(hostSuffix) tokens
-      # For more information see https://aka.ms/devspaces/routing
-      - $(spacePrefix)$(rootSpacePrefix)authapplication$(hostSuffix)
-configurations:
-  develop:
-    build:
-      dockerfile: Dockerfile.develop
-      useGitIgnore: true
-      args:
-        BUILD_CONFIGURATION: ${BUILD_CONFIGURATION:-Debug}
-    container:
-      sync:
-      - "**/Pages/**"
-      - "**/Views/**"
-      - "**/wwwroot/**"
-      - "!**/*.{sln,csproj}"
-      command: [dotnet, run, --no-restore, --no-build, --no-launch-profile, -c, "${BUILD_CONFIGURATION:-Debug}"]
-      iterate:
-        processesToKill: [dotnet, vsdbg, AuthApplication]
-        buildCommands:
-        - [dotnet, build, --no-restore, -c, "${BUILD_CONFIGURATION:-Debug}"]
