+++ src/UserApplication/.vscode/launch.json
@@ -18,7 +18,7 @@
                 "quoteArgs": false
             },
             "cwd": "/src",
-            "program": "/src/bin/Debug/netcoreapp2.1/AuthApplication.dll",
+            "program": "/src/bin/Debug/netcoreapp2.1/UserApplication.dll",
             "sourceFileMap": {
                 "/src": "${workspaceFolder}/."
             }
