+++ src/UserApplication/Dockerfile.develop
@@ -5,8 +5,8 @@ ENV DOTNET_USE_POLLING_FILE_WATCHER=true
 EXPOSE 80
 
 WORKDIR /src
-COPY ["AuthApplication.csproj", "./"]
-RUN dotnet restore "AuthApplication.csproj"
+COPY ["UserApplication.csproj", "./"]
+RUN dotnet restore "UserApplication.csproj"
 COPY . .
 RUN dotnet build --no-restore -c $BUILD_CONFIGURATION
