+++ src/UserApplication/Properties/launchSettings.json
@@ -17,7 +17,7 @@
         "ASPNETCORE_ENVIRONMENT": "Development"
       }
     },
-    "AuthApplication": {
+    "UserApplication": {
       "commandName": "Project",
       "launchBrowser": true,
       "launchUrl": "api/values",
