+++ src/UserApplication/appsettings.Development.json
