+++ src/UserApplication/azds.yaml
@@ -1,0 +1,51 @@
+kind: helm-release
+apiVersion: 1.1
+build:
+  context: .
+  dockerfile: Dockerfile
+install:
+  chart: charts/userapplication
+  values:
+    - values.dev.yaml?
+    - secrets.dev.yaml?
+  set:
+    replicaCount: 1
+    image:
+      repository: acmnotesregistry.azurecr.io/userapplication
+      tag: $(tag)
+      pullPolicy: Never
+    ingress:
+      annotations:
+        kubernetes.io/ingress.class: traefik-azds
+      hosts:
+        # This expands to form the service's public URL: [space.s.][rootSpace.]userapplication.<random suffix>.<region>.azds.io
+        # Customize the public URL by changing the 'userapplication' text between the $(rootSpacePrefix) and $(hostSuffix) tokens
+        # For more information see https://aka.ms/devspaces/routing
+        - $(spacePrefix)$(rootSpacePrefix)userapplication$(hostSuffix)
+configurations:
+  develop:
+    build:
+      dockerfile: Dockerfile.develop
+      useGitIgnore: true
+      args:
+        BUILD_CONFIGURATION: ${BUILD_CONFIGURATION:-Debug}
+    container:
+      sync:
+        - "**/Pages/**"
+        - "**/Views/**"
+        - "**/wwwroot/**"
+        - "!**/*.{sln,csproj}"
+      command:
+        [
+          dotnet,
+          run,
+          --no-restore,
+          --no-build,
+          --no-launch-profile,
+          -c,
+          "${BUILD_CONFIGURATION:-Debug}",
+        ]
+      iterate:
+        processesToKill: [dotnet, vsdbg, UserApplication]
+        buildCommands:
+          - [dotnet, build, --no-restore, -c, "${BUILD_CONFIGURATION:-Debug}"]
