+++ src/UserApplication/charts/authapplication/templates/ingress.yaml
@@ -1,5 +1,5 @@
 {{- if .Values.ingress.enabled -}}
-{{- $fullName := include "authapplication.fullname" . -}}
+{{- $fullName := include "userapplication.fullname" . -}}
 {{- $servicePort := .Values.service.port -}}
 {{- $ingressPath := .Values.ingress.path -}}
 apiVersion: extensions/v1beta1
@@ -7,8 +7,8 @@ kind: Ingress
 metadata:
   name: {{ $fullName }}
   labels:
-    app: {{ template "authapplication.name" . }}
-    chart: {{ template "authapplication.chart" . }}
+    app: {{ template "userapplication.name" . }}
+    chart: {{ template "userapplication.chart" . }}
     release: {{ .Release.Name }}
     heritage: {{ .Release.Service }}
 {{- with .Values.ingress.annotations }}
