+++ src/UserApplication/charts/authapplication/templates/secrets.yaml
@@ -3,7 +3,7 @@
 apiVersion: v1
 kind: Secret
 metadata:
-  name: {{ template "authapplication.fullname" $root }}-{{ $name | lower }}
+  name: {{ template "userapplication.fullname" $root }}-{{ $name | lower }}
 data:
   {{- range $key, $value := $values }}
   {{ $key }}: {{ $value | b64enc }}
