+++ src/UserApplication/charts/authapplication/templates/service.yaml
@@ -1,10 +1,10 @@
 apiVersion: v1
 kind: Service
 metadata:
-  name: {{ template "authapplication.fullname" . }}
+  name: {{ template "userapplication.fullname" . }}
   labels:
-    app: {{ template "authapplication.name" . }}
-    chart: {{ template "authapplication.chart" . }}
+    app: {{ template "userapplication.name" . }}
+    chart: {{ template "userapplication.chart" . }}
     release: {{ .Release.Name }}
     heritage: {{ .Release.Service }}
 spec:
@@ -15,5 +15,5 @@ spec:
       protocol: TCP
       name: http
   selector:
-    app: {{ template "authapplication.name" . }}
+    app: {{ template "userapplication.name" . }}
     release: {{ .Release.Name }}
