+++ src/UserApplication/charts/authapplication/values.yaml
@@ -1,10 +1,10 @@
-# Default values for authapplication.
+# Default values for userapplication.
 # This is a YAML-formatted file.
 # Declare variables to be passed into your templates.
-fullnameOverride: authapplication
+fullnameOverride: userapplication
 replicaCount: 1
 image:
-  repository: acmnotesregistry.azurecr.io/authapplication
+  repository: acmnotesregistry.azurecr.io/userapplication
   tag: stable
   pullPolicy: IfNotPresent
 imagePullSecrets: []
@@ -27,7 +27,7 @@ ingress:
     kubernetes.io/ingress.class: traefik-azds
   path: /api/auth
   hosts:
-    - dev.authapplication.7ssmwvmdd8.wus2.azds.io
+    - dev.userapplication.7ssmwvmdd8.wus2.azds.io
   tls: []
 secrets: {}
   # Optionally specify a set of secret objects whose values
